Break the webpack configuration into smaller plugins
