+++ package.json
@@ -69,6 +69,7 @@
     "url-loader": "^0.5.6",
     "webpack": "^1.9.6",
     "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.0.0"
+    "webpack-hot-middleware": "^2.0.0",
+    "webpack-merge": "^0.7.3"
   }
 }
